extract FizzyButton into its own component
